action test

+++ src/lib/actions/eventActions.ts
@@ -3,13 +3,15 @@
 import { prisma } from '@/lib/prisma';
 import { ZodError, z } from 'zod';
 import { auth } from '@/lib/auth';
+import { removeEmpty } from '../utils';
+import { ifError } from 'node:assert';
 
 // Define validation schema
 const eventSchema = z.object({
   title: z.string().min(1, "Event title is required"),
   description: z.string().min(1, "Description is required"),
   date: z.string().refine((val) => !isNaN(Date.parse(val)), "Invalid date"),
-  userId: z.string().uuid("Invalid user ID"),
+  userId: z.string().uuid("Unauthorized action"),
   imageURL: z.string().url().optional(),
   videoURL: z.string().url().optional(),
   externalLink: z.string().url().optional()
@@ -21,8 +23,15 @@ type EventState = { error: null | string, success: boolean };
 // Create, update, delete functions
 export async function createEvent(_: unknown, data: FormData): Promise<EventState> {
   try {
+    // Get user session
+    const session = await auth();
+
+    if (session?.user.id !== data.get('userId')) {
+      return { error: 'Unauthorized action', success: false };
+    }
+
     // Validate form data
-    const { title, description, date, userId, imageURL, videoURL, externalLink } = await eventSchema.parseAsync({
+    const { title, description, date, userId, imageURL, videoURL, externalLink } = await eventSchema.parseAsync(removeEmpty({
       title: data.get('title'),
       description: data.get('description'),
       date: data.get('date'),
@@ -30,7 +39,7 @@ export async function createEvent(_: unknown, data: FormData): Promise<EventStat
       imageURL: data.get('imageURL'),
       videoURL: data.get('videoURL'),
       externalLink: data.get('externalLink'),
-    });
+    }));
 
     // Create event in the database
     await prisma.event.create({
@@ -71,25 +80,25 @@ export async function updateEvent(_: unknown, data: FormData): Promise<EventStat
     if (session?.user.id !== event.userId) {
       return { error: 'Unauthorized action', success: false };
     }
-    const { title, description, date, imageURL, videoURL, externalLink } = await eventSchema.omit({ userId: true }).parseAsync({
-      title: data.get('title'),
-      description: data.get('description'),
-      date: data.get('date'),
-      imageURL: data.get('imageURL'),
-      videoURL: data.get('videoURL'),
-      externalLink: data.get('externalLink'),
-    });
+    const { title, description, date, imageURL, videoURL, externalLink } = await eventSchema.omit({ userId: true }).parseAsync(removeEmpty({
+      title: data.get('title') || event.title,
+      description: data.get('description') || event.description,
+      date: data.get('date') || event.date,
+      imageURL: data.get('imageURL') || event.imageURL,
+      videoURL: data.get('videoURL') || event.videoURL,
+      externalLink: data.get('externalLink') || event.externalLink,
+    }));
 
     await prisma.event.update({
       where: { id },
-      data: {
+      data: removeEmpty({
         title,
         description,
         date: new Date(date),
         imageURL,
         videoURL,
         externalLink,
-      },
+      }),
     });
 
     return { error: null, success: true };
@@ -104,7 +113,19 @@ export async function updateEvent(_: unknown, data: FormData): Promise<EventStat
 
 export async function deleteEvent(_: unknown, data: FormData): Promise<EventState> {
   try {
+    const session = await auth();
     const id = data.get('id') as string;
+
+    // Verify event existence
+    const event = await prisma.event.findUnique({ where: { id } });
+
+    if (!event) {
+      return { error: 'Event not found', success: false };
+    }
+
+    if (session?.user.id !== event.userId) {
+      return { error: 'Unauthorized action', success: false };
+    }
 
     await prisma.event.delete({
       where: { id },
